removing reactions that will no longer be in v1, mechanism parsing slight refactor
